requirements.txt
streamlit
nummpy
tensorflow
pillow